Update docs to reflect changes in API
Added a lot of CTA fields for scheduling as well as lastUsedEmpwoerMts and currentCtaId. 

RegionIds within the CTA Object might need some clarification, if the region has a new CTA attached to it, will it become [] for the previous CTA? Noticing there's no regionIds column in the Campaign_activities table and the default_activity_id is still in the regions table and overwrites each time.